[CPU] refine CPU integration code (#7647)
